rename inference -> generate
