+++ mlora/__init__.py
@@ -6,7 +6,7 @@ from mlora.model_llama import LlamaModel
 from mlora.model_chatglm import ChatGLMModel
 from mlora.modelargs import LLMModelArgs, MultiLoraBatchData, LoraBatchDataConfig, LoraConfig, MixConfig
 from mlora.dispatcher import TrainTask, Dispatcher
-from mlora.inference import GenerateConfig, generate
+from mlora.generate import GenerateConfig, generate
 
 __all__ = [
     "convert_hf_to_pth",
